fixed your gulp stylus bug
the problem was that you were trying to compile each one of the styl files individually when your files were set up to be compiled in a cascading fashion, compiling one file which imports the rest and brings them into the scope. when it tried to compile them all individually it was running into variable not defined issues.